requirements.txt
fastapi==0.114.1
langchain==0.2.16
langchain_community==0.2.16
langchain_core==0.2.39
langchain_openai==0.1.24
langgraph==0.2.19
Pillow==10.4.0
pydantic==2.9.1
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
setuptools==65.5.0
SQLAlchemy==2.0.32
streamlit==1.37.1
uvicorn==0.30.6
xinference==0.14.4
